✨ feat: adicionar componentes de ranking e cálculo de média de votos
