+++ src/app/home/types.ts
@@ -1,0 +1,7 @@
+export interface Jogador {
+  id: string;
+  nome: string;
+  assistencias: number;
+  gols: number;
+  votos: { userId: string; vote: number }[];
+}
